refactor: NotificationSettingStorage interface 추가
