+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -92,7 +92,7 @@ class MainActivity : ComponentActivity() {
                                     cartRepository = app.cartRepository,
                                     recentlyViewedProductRepository = app.recentlyViewedProductRepository,
                                     productRepository = app.productRepository,
-                                    notificationStorage = ShoppingApplication.notificationSetting,
+                                    notificationSettingStorage = ShoppingApplication.notificationSetting,
                                 )
                             )
                             ShoppingRoute(
